+++ tests/data/goldens/typing_pathlib_merge_0001.py
@@ -1,4 +1,2 @@
-
-
 class A:
     pass